+++ src/components/currency.js
@@ -65,6 +65,7 @@ module.exports = function(app) {
       $templateCache.put('formio/components/currency/validate.html',
         '<ng-form>' +
           '<form-builder-option property="validate.required"></form-builder-option>' +
+          '<form-builder-option property="validate.customMessage"></form-builder-option>' +
           '<form-builder-option-custom-validation></form-builder-option-custom-validation>' +
         '</ng-form>'
       );