+++ src/ngFormBuilder.js
@@ -97,6 +97,8 @@ app.directive('formBuilderOptionLabelPosition', require('./directives/formBuilde
 
 app.directive('formBuilderOptionOptionsLabelPosition', require('./directives/formBuilderOptionOptionsLabelPosition'));
 
+app.directive('formBuilderOptionShortcut', require('./directives/formBuilderOptionShortcut'));
+
 app.directive('formBuilderOptionTags', require('./directives/formBuilderOptionTags'));
 
 app.directive('validApiKey', require('./directives/validApiKey'));
@@ -106,6 +108,8 @@ app.directive('formBuilderOptionCustomValidation', require('./directives/formBui
 app.directive('formBuilderTooltip', require('./directives/formBuilderTooltip'));
 
 app.directive('valueBuilder', require('./directives/valueBuilder'));
+
+app.directive('valueBuilderWithShortcuts', require('./directives/valueBuilderWithShortcuts'));
 
 app.directive('objectBuilder', require('./directives/objectBuilder'));